game logic
i need to add a new layer, remove the main and make it a package.
when i do that and make it fully operable via a fully efficient system than i can make a model simple model that plays the game

maybe add a stepping system idk how to yet tho.
so for now i need to add a split and a blackjack win system

if i have that step completed than ill be able to add a simple top layer that carries out commands based on a model. loop it mutate it itterate it and than disciminate on that.